+++ src/server/routes/cnf/[oneTimeLink].ts
@@ -30,8 +30,8 @@ export default defineEventHandler(async (event) => {
   setHeader(
     event,
     'Content-Disposition',
-    `attachment; filename="${client.name}.conf"`
+    `attachment; filename="${WireGuard.cleanClientFilename(client.name) || client.id}.conf"`
   );
-  setHeader(event, 'Content-Type', 'text/plain');
+  setHeader(event, 'Content-Type', 'application/octet-stream');
   return config;
 });